pyproject.toml
[project]
authors = [
    {name = "Sergio Mattei", email = "sergiomattei@outlook.com"},
]
requires-python = "<4.0,>=3.9"
dependencies = [
    "PyJWT<3.0.0,>=2.8.0",
    "click<9.0.0,>=8.1.7",
    "rich<14.0.0,>=13.7.0",
    "requests<3.0.0,>=2.31.0",
    "cryptography<42.0.0,>=41.0.7",
    "flask<4.0.0,>=3.0.0",
]
name = "am-keyman"
version = "0.1.5"
description = ""
readme = "README.md"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
am_keyman = ["*.html"]

[project.scripts]
am-keyman = "am_keyman.main:cli"
